Added a monitor that follows the log, printing out verified releases.
This monitor persists a checkpoint locally in order to maintain a consistent view of the log's history between runs. For each leaf:
 * The raw data is downloaded into memory
 * The leaf is verified to be included within the golden checkpoint
 * XXX: The trailing newlines are removed from the leaf data
 * The leaf data is parsed as a note, and confirmed to be signed by the expected key
 * The note text is parsed as a FirmwareRelease
 * Some useful subset of data from the verified release is printed to the log

The monitor will detect new checkpoints by polling the log periodically.
Any error will cause the tool to fail; we may wish to make some failures print warnings/errors in the log instead.
On succesfully verifying all leaves up to a checkpoint size, the checkpoint is persisted locally so that the visited leaves will not be visited again.